+++ opts/single_axis.go
@@ -1,0 +1,51 @@
+package opts
+
+// SingleAxis is the option set for single axis.
+// https://echarts.apache.org/en/option.html#singleAxis
+type SingleAxis struct {
+	// The minimum value of axis.
+	// It can be set to a special value 'dataMin' so that the minimum value on this axis is set to be the minimum label.
+	// It will be automatically computed to make sure axis tick is equally distributed when not set.
+	Min interface{} `json:"min,omitempty"`
+
+	// The maximum value of axis.
+	// It can be set to a special value 'dataMax' so that the minimum value on this axis is set to be the maximum label.
+	// It will be automatically computed to make sure axis tick is equally distributed when not set.
+	Max interface{} `json:"max,omitempty"`
+
+	// Type of axis.
+	// Option:
+	// * 'value': Numerical axis, suitable for continuous data.
+	// * 'category': Category axis, suitable for discrete category data.
+	//   Category data can be auto retrieved from series.data or dataset.source,
+	//   or can be specified via xAxis.data.
+	// * 'time' Time axis, suitable for continuous time series data. As compared to value axis,
+	//   it has a better formatting for time and a different tick calculation method. For example,
+	//   it decides to use month, week, day or hour for tick based on the range of span.
+	// * 'log' Log axis, suitable for log data.
+	Type string `json:"type,omitempty"`
+
+	// Distance between grid component and the left side of the container.
+	// left value can be instant pixel value like 20; it can also be a percentage
+	// value relative to container width like '20%'; and it can also be 'left', 'center', or 'right'.
+	// If the left value is set to be 'left', 'center', or 'right',
+	// then the component will be aligned automatically based on position.
+	Left string `json:"left,omitempty"`
+
+	// Distance between grid component and the right side of the container.
+	// right value can be instant pixel value like 20; it can also be a percentage
+	// value relative to container width like '20%'.
+	Right string `json:"right,omitempty"`
+
+	// Distance between grid component and the top side of the container.
+	// top value can be instant pixel value like 20; it can also be a percentage
+	// value relative to container width like '20%'; and it can also be 'top', 'middle', or 'bottom'.
+	// If the left value is set to be 'top', 'middle', or 'bottom',
+	// then the component will be aligned automatically based on position.
+	Top string `json:"top,omitempty"`
+
+	// Distance between grid component and the bottom side of the container.
+	// bottom value can be instant pixel value like 20; it can also be a percentage
+	// value relative to container width like '20%'.
+	Bottom string `json:"bottom,omitempty"`
+}
